requirements.txt
pycryptodome>=3.15.0
flask>=2.0.0
pyinstaller>=5.0.0
psutil>=5.9.0
pefile>=2022.5.30
capstone>=4.0.2
colorama>=0.4.5
click>=8.0.0
rich>=12.0.0
PyYAML>=6.0
requests>=2.28.0 